使用LocalKeyManager管理本地密钥对
